refactor(sentry): split the handled-family lookup out of the autofix queue I/O
scripts/sentry-autofix-queue-io.mjs sat at 583 of the repo's 600-line soft cap —
under it, but with 17 lines of headroom, which is not enough to survive the next
change to the selection leg.

Move `listHandledShortIds` and `MAX_HANDLED_ID_QUERIES` to
scripts/sentry-autofix-family-handled.mjs: 583 -> 428, new module 197. The
dependency runs one way (new module -> queue-io) with no re-export shim, so no
ESM cycle is reachable.

Pure refactor, proven rather than asserted: the selector CLI was run across ten
paths against a stub `gh` and diffed against `git archive origin/main` of the
pre-split tree — stdout, stderr, exit codes and all four report files are
byte-identical.

Also closes two enforcement gaps the split exposed: the new module carries
MAX_HANDLED_ID_QUERIES, a term in the finalize suite's select-job timeout pin,
so it routes to that suite as well as the select suite; and the selection leg
now has a machine-enforced 600-line pin, which it did not have before.
